thrp: fixed some warnings

+++ src/libromdata/TouhouReplay.cpp
@@ -87,7 +87,7 @@ namespace LibRomData {
 
 		// First, read an unencrypted area as-is, and advance ptr, size and cpos
 		if (cpos < offset) {
-			count = file->read(ptr, std::min<uint64_t>(offset, cpos + size) - cpos);
+			count = file->read(ptr, (size_t)(std::min<uint64_t>(offset, cpos + size) - cpos));
 			*(char**)&ptr += count;
 			cpos += count;
 			size -= count;
@@ -98,7 +98,7 @@ namespace LibRomData {
 			count2 = file->read(ptr, size);
 			
 			uint8_t k = (uint8_t)( key + 7 * (cpos - offset) ); // calculate key for our file position
-			for (int i = 0; i<count2; i++) {
+			for (size_t i = 0; i<count2; i++) {
 				((uint8_t*)ptr)[i] -= k;
 				k += 7;
 			}
